windows build fix

+++ aui.audio/src/AUI/Audio/Platform/win32/DirectSoundAudioPlayer.cpp
@@ -137,7 +137,7 @@ void DirectSoundAudioPlayer::setupSecondaryBuffer() {
     waveFormat.cbSize = sizeof(waveFormat);
     waveFormat.wFormatTag = WAVE_FORMAT_PCM;
     waveFormat.wBitsPerSample = info.bitsPerSample();
-    waveFormat.nChannels = info.channelCount;
+    waveFormat.nChannels = static_cast<WORD>(info.channelCount);
     waveFormat.nSamplesPerSec = info.sampleRate;
     waveFormat.nBlockAlign = waveFormat.nChannels * (waveFormat.wBitsPerSample / 8);
